a a a aa a

+++ PolyFinancial-TXJ-html/js/ctrl/mine/setting.js
@@ -16,23 +16,21 @@ app.controller("settingCtrl",function ($scope,$http,$state,serviceHTTP,$statePar
             },
             callback: function(result) {
                 if(result === true){
-                    // serviceHTTP.loginOutHTTP(id).then(function successCallback(response) {
-                    //     // 请求成功执行代码
-                    //     console.log(response);
-                    //     if(response.data.message === "success") {
-                    //         bootbox.alert("删除成功");
-                    //         $state.reload('backStage.debt');
-                    //     }
-                    //     else {
-                    //         bootbox.alert(response.data.message);
-                    //     }
-                    // }, function errorCallback(res) {
-                    //     // 请求失败执行代码
-                    // });
-                    bootbox.alert("您已经退出登录");
+                    serviceHTTP.logoutHTTP().then(function successCallback(response) {
+                        // 请求成功执行代码
+                        console.log(response);
+                        if(response.data.code == 0) {
+                            bootbox.alert("您已经退出登录");
+                            sessionStorage.clear();
+                            $state.go('home.mine')
+                        }
+                        else {
+                            bootbox.alert(response.data.message);
+                        }
+                    }, function errorCallback(res) {
+                        // 请求失败执行代码
+                    });
 
-                    sessionStorage.clear();
-                    $state.go('home.mine')
 
 
                 }
@@ -47,6 +45,5 @@ app.controller("settingCtrl",function ($scope,$http,$state,serviceHTTP,$statePar
             return false
         }
         // window.history.back(-1);
-        $state.go('home');
     }
 });
